Conflicts:
	src/acc/acc_helper_functions_impl.h
	src/apps/image_handler.cpp
	src/jaz/ctf/ctf_refiner.cpp
	src/jaz/motion/motion_refiner.cpp
	src/jaz/reference_map.cpp
	src/macros.h
	src/pipeline_jobs.cpp
	src/pipeline_jobs.h

+++ src/acc/acc_helper_functions_impl.h
@@ -686,7 +686,7 @@ void runDiff2KernelCoarse(
 				CRITICAL(ERR_TRANSLIM);
 #endif
 
-			unsigned rest = orientation_num % blocks3D;
+			long unsigned rest = orientation_num % blocks3D;
 			long unsigned even_orientation_num = orientation_num - rest;
 // TODO - find a more compact way to represent these combinations resulting in
 // a single call to diff2_course?
@@ -696,7 +696,7 @@ void runDiff2KernelCoarse(
 				{
 					if(data_is_3D)
 						AccUtilities::diff2_coarse<true,true, D2C_BLOCK_SIZE_DATA3D, D2C_EULERS_PER_BLOCK_DATA3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_DATA3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_DATA3D, 
 							D2C_BLOCK_SIZE_DATA3D,
 							d_eulers,
 							trans_x,
@@ -712,7 +712,7 @@ void runDiff2KernelCoarse(
 							stream);
 					else
 						AccUtilities::diff2_coarse<true,false, D2C_BLOCK_SIZE_REF3D, D2C_EULERS_PER_BLOCK_REF3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_REF3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_REF3D, 
 							D2C_BLOCK_SIZE_REF3D,
 							d_eulers,
 							trans_x,
@@ -770,7 +770,7 @@ void runDiff2KernelCoarse(
 				{
 					if(data_is_3D)
 						AccUtilities::diff2_coarse<true,true, D2C_BLOCK_SIZE_DATA3D*2, D2C_EULERS_PER_BLOCK_DATA3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_DATA3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_DATA3D,
 							D2C_BLOCK_SIZE_DATA3D*2,
 							d_eulers,
 							trans_x,
@@ -786,7 +786,7 @@ void runDiff2KernelCoarse(
 							stream);
 					else
 						AccUtilities::diff2_coarse<true,false, D2C_BLOCK_SIZE_REF3D*2, D2C_EULERS_PER_BLOCK_REF3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_REF3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_REF3D,
 							D2C_BLOCK_SIZE_REF3D*2,
 							d_eulers,
 							trans_x,
@@ -845,7 +845,7 @@ void runDiff2KernelCoarse(
 				{
 					if(data_is_3D)
 						AccUtilities::diff2_coarse<true,true, D2C_BLOCK_SIZE_DATA3D*4, D2C_EULERS_PER_BLOCK_DATA3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_DATA3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_DATA3D,
 							D2C_BLOCK_SIZE_DATA3D*4,
 							d_eulers,
 							trans_x,
@@ -861,7 +861,7 @@ void runDiff2KernelCoarse(
 							stream);
 					else
 						AccUtilities::diff2_coarse<true,false, D2C_BLOCK_SIZE_REF3D*4, D2C_EULERS_PER_BLOCK_REF3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_REF3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_REF3D,
 							D2C_BLOCK_SIZE_REF3D*4,
 							d_eulers,
 							trans_x,
@@ -920,7 +920,7 @@ void runDiff2KernelCoarse(
 				{
 					if(data_is_3D)
 						AccUtilities::diff2_coarse<true,true, D2C_BLOCK_SIZE_DATA3D*8, D2C_EULERS_PER_BLOCK_DATA3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_DATA3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_DATA3D,
 							D2C_BLOCK_SIZE_DATA3D*8,
 							d_eulers,
 							trans_x,
@@ -936,7 +936,7 @@ void runDiff2KernelCoarse(
 							stream);
 					else
 						AccUtilities::diff2_coarse<true,false, D2C_BLOCK_SIZE_REF3D*8, D2C_EULERS_PER_BLOCK_REF3D, 4>(
-							even_orientation_num/D2C_EULERS_PER_BLOCK_REF3D,
+							even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_REF3D,
 							D2C_BLOCK_SIZE_REF3D*8,
 							d_eulers,
 							trans_x,
@@ -1001,14 +1001,14 @@ void runDiff2KernelCoarse(
 			}
 
 
-			unsigned rest = orientation_num % D2C_EULERS_PER_BLOCK_2D;
+			long unsigned rest = orientation_num % (unsigned long)D2C_EULERS_PER_BLOCK_2D;
 			long unsigned even_orientation_num = orientation_num - rest;
 
 			if (even_orientation_num != 0)
 			{
 				if(data_is_3D)
 					AccUtilities::diff2_coarse<false,true, D2C_BLOCK_SIZE_2D, D2C_EULERS_PER_BLOCK_2D, 2>(
-						even_orientation_num/D2C_EULERS_PER_BLOCK_2D,
+						even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_2D,
 						D2C_BLOCK_SIZE_2D,
 						d_eulers,
 						trans_x,
@@ -1024,7 +1024,7 @@ void runDiff2KernelCoarse(
 						stream);
 				else
 					AccUtilities::diff2_coarse<false,false, D2C_BLOCK_SIZE_2D, D2C_EULERS_PER_BLOCK_2D, 2>(
-						even_orientation_num/D2C_EULERS_PER_BLOCK_2D,
+						even_orientation_num/(unsigned long)D2C_EULERS_PER_BLOCK_2D,
 						D2C_BLOCK_SIZE_2D,
 						d_eulers,
 						trans_x,
@@ -1168,7 +1168,7 @@ void runDiff2KernelFine(
 		bool do_CC,
 		bool data_is_3D)
 {
-    int block_dim = job_num_count;
+    long unsigned block_dim = job_num_count;
 
     if(!do_CC)
     {
